% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abb.R
\docType{data}
\name{abb}
\alias{abb}
\title{Abbreviations matrix}
\format{A \code{matrix} object with 56027 rows and 3 columns (word, abbreviations, languages)}
\usage{
abb
}
\description{
Abbreviations matrix from ISSN LTWA
}
\details{
Original abbrevations are from http://www.issn.org/wp-content/uploads/2013/07/LTWA_20150724.pdf

The matrix was parsed from the website using \code{\link{get_abbrev}} and is used in \code{\link{abbrev}}.
}
\keyword{datasets}
