% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrev.R
\name{abbrev}
\alias{abbrev}
\title{Get the abbreviated journal title}
\usage{
abbrev(title, rems = c("and", "of", "\\\\&", "the", "on"))
}
\arguments{
\item{title}{chr string of the journal name}

\item{rems}{chr string of prepositions or other words to remove from the journal title}
}
\value{
A chr string for the abbreviated journal title
}
\description{
Get the abbreviated journal title
}
\examples{
abbrev('Journal of Applied Mathematics')
abbrev('Fundamental and Applied Limnology')
abbrev('Science')
}

