% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbl_ext.R
\name{bbl_ext}
\alias{bbl_ext}
\title{Extract bib entries form a bib file that are in bbl}
\usage{
bbl_ext(bbl_fl = "inst/manu.bbl", bib_fl = "inst/refs_all.bib",
  bib_new = NULL)
}
\arguments{
\item{bbl_fl}{chr string of bbl file with entries to extract}

\item{bib_fl}{chr string of bib file for all entries}

\item{bib_new}{chr string of path for new bib file with abbreviated titles, appends '_kp' to input file name if blank}
}
\value{
A new bib file is saved, nothing is returned to the console
}
\description{
Extract bib entries form a bib file that are in bbl
}
\details{
This function assumes that the bib file with entries to extract based on keys in bbl includes extra entries not needed for a document.
}
\examples{
bbl_ext('inst/manu.bbl', 'inst/refs_all.bib')
}

