% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_scrp.R
\name{bib_scrp}
\alias{bib_scrp}
\title{Find bibtex tags in a document, scrape entries from external bib file, save new}
\usage{
bib_scrp(
  rmd_in,
  bib_new = "refs.bib",
  ext_bib = "https://fawda123.github.io/refs/refs.bib"
)
}
\arguments{
\item{rmd_in}{chr string of path to Rmd file}

\item{bib_new}{path for output .bib file}

\item{ext_bib}{chr string of url for external bib file to scrape}
}
\value{
A new bib file is saved, nothing is returned to the console
}
\description{
Find bibtex tags in a document, scrape entries from external bib file, save new
}
\details{
Currently only works for .Rmd files with the Rmarkdown format of bibtex tags, e.g., \code{[@Jones18]}.  The user must also check the external bib file to make sure all tags in the .Rmd file are present.
}
\examples{
\dontrun{
bib_scrp('document.Rmd', 'refs.bib')
}
}
