% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_shr.R
\name{bib_shr}
\alias{bib_shr}
\title{Convert long journal names in a bib file to abbreviated}
\usage{
bib_shr(bib_fl, bib_new = NULL, trace = TRUE)
}
\arguments{
\item{bib_fl}{chr string of path to bib file}

\item{bib_new}{chr string of path for new bib file with abbreviated titles, appends '_new' to input file name if blank}

\item{trace}{logical to print progress to console}
}
\value{
A new bib file is saved, nothing is returned to the console
}
\description{
Convert long journal names in a bib file to abbreviated
}
\examples{
bib_shr('inst/refs2')
}

