% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capwords.R
\name{capwords}
\alias{capwords}
\title{Convert a character string to title case}
\usage{
capwords(s, strict = FALSE)
}
\arguments{
\item{s}{chr string for title case}

\item{strict}{logical}
}
\value{
A chr string in title case
}
\description{
Convert a character string to title case
}
\details{
This function is from the help file for \code{\link[base]{toupper}}
}

