% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abbrev.R
\name{get_abbrev}
\alias{get_abbrev}
\title{Get ISSN LTWA}
\usage{

  get_abbrev(url = "http://www.issn.org/wp-content/uploads/2013/07/LTWA_20150724.pdf")
}
\arguments{
\item{url}{chr string of pdf to pars}
}
\value{
A three-column matrix with the word, abbreviation, and language
}
\description{
Get the ISSN List of Title Word Abbreviations
}
\details{
The LTWA pdf is parsed from the online URL
}
\examples{
get_abbrev()
}

